Moves OTP element.
