+++ link/src/androidTest/java/com/stripe/android/link/ui/verification/VerificationScreenTest.kt
@@ -24,8 +24,8 @@ import com.stripe.android.link.R
 import com.stripe.android.link.theme.DefaultLinkTheme
 import com.stripe.android.link.ui.ErrorMessage
 import com.stripe.android.uicore.elements.IdentifierSpec
-import com.stripe.android.ui.core.elements.OTPController
-import com.stripe.android.ui.core.elements.OTPElement
+import com.stripe.android.uicore.elements.OTPController
+import com.stripe.android.uicore.elements.OTPElement
 import com.stripe.android.ui.core.elements.OTPSpec
 import com.stripe.android.uicore.forms.FormFieldEntry
 import org.junit.Rule
