+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/FormUI.kt
@@ -19,18 +19,18 @@ import com.stripe.android.ui.core.elements.BsbElementUI
 import com.stripe.android.ui.core.elements.CardDetailsSectionElement
 import com.stripe.android.ui.core.elements.CardDetailsSectionElementUI
 import com.stripe.android.ui.core.elements.EmptyFormElement
-import com.stripe.android.ui.core.elements.FormElement
 import com.stripe.android.ui.core.elements.MandateTextElement
 import com.stripe.android.ui.core.elements.MandateTextUI
-import com.stripe.android.ui.core.elements.OTPElement
-import com.stripe.android.ui.core.elements.OTPElementUI
 import com.stripe.android.ui.core.elements.SaveForFutureUseElement
 import com.stripe.android.ui.core.elements.SaveForFutureUseElementUI
 import com.stripe.android.ui.core.elements.SectionElement
 import com.stripe.android.ui.core.elements.SectionElementUI
 import com.stripe.android.ui.core.elements.StaticTextElement
 import com.stripe.android.ui.core.elements.StaticTextElementUI
+import com.stripe.android.uicore.elements.FormElement
 import com.stripe.android.uicore.elements.IdentifierSpec
+import com.stripe.android.uicore.elements.OTPElement
+import com.stripe.android.uicore.elements.OTPElementUI
 import kotlinx.coroutines.flow.Flow
 
 @Composable
