+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/AffirmTextSpec.kt
@@ -1,5 +1,6 @@
 package com.stripe.android.ui.core.elements
 
+import com.stripe.android.uicore.elements.FormElement
 import com.stripe.android.uicore.elements.IdentifierSpec
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
