--- payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/FormElement.kt
@@ -1,22 +1,0 @@
-package com.stripe.android.ui.core.elements
-
-import androidx.annotation.RestrictTo
-import com.stripe.android.uicore.elements.Controller
-import com.stripe.android.uicore.elements.IdentifierSpec
-import com.stripe.android.uicore.forms.FormFieldEntry
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-
-/**
- * This is used to define each section in the visual form layout.
- * Each item in the layout has an identifier and a controller associated with it.
- */
-@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-sealed class FormElement {
-    abstract val identifier: IdentifierSpec
-    abstract val controller: Controller?
-
-    abstract fun getFormFieldValueFlow(): Flow<List<Pair<IdentifierSpec, FormFieldEntry>>>
-    open fun getTextFieldIdentifiers(): Flow<List<IdentifierSpec>> =
-        MutableStateFlow(emptyList())
-}
