+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/MandateTextElement.kt
@@ -1,6 +1,7 @@
 package com.stripe.android.ui.core.elements
 
 import androidx.annotation.RestrictTo
+import com.stripe.android.uicore.elements.FormElement
 import com.stripe.android.uicore.elements.IdentifierSpec
 import com.stripe.android.uicore.elements.InputController
 import com.stripe.android.uicore.forms.FormFieldEntry
@@ -13,7 +14,7 @@ data class MandateTextElement(
     val stringResId: Int,
     val merchantName: String?,
     override val controller: InputController? = null
-) : FormElement() {
+) : FormElement {
     override fun getFormFieldValueFlow(): Flow<List<Pair<IdentifierSpec, FormFieldEntry>>> =
         MutableStateFlow(emptyList())
 }
