+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/MandateTextSpec.kt
@@ -2,6 +2,7 @@ package com.stripe.android.ui.core.elements
 
 import androidx.annotation.RestrictTo
 import androidx.annotation.StringRes
+import com.stripe.android.uicore.elements.FormElement
 import com.stripe.android.uicore.elements.IdentifierSpec
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
