+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/OTPSpec.kt
@@ -2,6 +2,8 @@ package com.stripe.android.ui.core.elements
 
 import androidx.annotation.RestrictTo
 import com.stripe.android.uicore.elements.IdentifierSpec
+import com.stripe.android.uicore.elements.OTPController
+import com.stripe.android.uicore.elements.OTPElement
 import kotlinx.serialization.Serializable
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
