+++ stripe-ui-core/src/main/java/com/stripe/android/uicore/elements/OTPController.kt
@@ -1,8 +1,7 @@
-package com.stripe.android.ui.core.elements
+package com.stripe.android.uicore.elements
 
 import androidx.annotation.RestrictTo
 import androidx.compose.ui.text.input.KeyboardType
-import com.stripe.android.uicore.elements.Controller
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.combine
